CORDA-3761: Move obtainIdentities() to KeyStoreHandler
