+++ node/src/test/kotlin/net/corda/node/internal/KeyStoreHandlerTest.kt
@@ -23,13 +23,13 @@ import java.io.IOException
 import java.security.KeyStoreException
 import java.security.cert.X509Certificate
 
-class NodeKeyStoreUtilitiesTest {
+class KeyStoreHandlerTest {
     @Test(timeout = 300_000)
     fun `initializing key store in non-dev mode with no key store`() {
         whenever(signingSupplier.get()).doAnswer { throw IOException() }
 
         assertThatThrownBy {
-            config.initKeyStores(cryptoService)
+            keyStoreHandler.initKeyStores()
         }.hasMessageContaining("One or more keyStores (identity or TLS) or trustStore not found.")
     }
 
@@ -38,7 +38,7 @@ class NodeKeyStoreUtilitiesTest {
         whenever(signingSupplier.get()).doAnswer { throw KeyStoreException() }
 
         assertThatThrownBy {
-            config.initKeyStores(cryptoService)
+            keyStoreHandler.initKeyStores()
         }.hasMessageContaining("At least one of the keystores or truststore passwords does not match configuration")
     }
 
@@ -48,7 +48,7 @@ class NodeKeyStoreUtilitiesTest {
         whenever(trustStore.iterator()).thenReturn(mock())
 
         assertThatThrownBy {
-            config.initKeyStores(cryptoService)
+            keyStoreHandler.initKeyStores()
         }.hasMessageContaining("Alias for trustRoot key not found. Please ensure you have an updated trustStore file")
     }
 
@@ -57,7 +57,7 @@ class NodeKeyStoreUtilitiesTest {
         whenever(keyStore.contains(CORDA_CLIENT_TLS)).thenReturn(false)
 
         assertThatThrownBy {
-            config.initKeyStores(cryptoService)
+            keyStoreHandler.initKeyStores()
         }.hasMessageContaining("Alias for TLS key not found. Please ensure you have an updated TLS keyStore file")
     }
 
@@ -67,13 +67,13 @@ class NodeKeyStoreUtilitiesTest {
         whenever(keyStore.query(any<X509KeyStore.() -> List<X509Certificate>>())).thenReturn(mutableListOf(untrustedRoot))
 
         assertThatThrownBy {
-            config.initKeyStores(cryptoService)
+            keyStoreHandler.initKeyStores()
         }.hasMessageContaining("TLS certificate must chain to the trusted root")
     }
 
     @Test(timeout = 300_000)
     fun `initializing key store should return valid certificate if certificate is valid`() {
-        val certificate = config.initKeyStores(cryptoService)
+        val certificate = keyStoreHandler.initKeyStores()
 
         assertThat(certificate).isEqualTo(listOf(trustRoot))
     }
@@ -87,7 +87,7 @@ class NodeKeyStoreUtilitiesTest {
         whenever(keySupplier.getOptional()).thenReturn(mock())
         whenever(signingSupplier.getOptional()).thenReturn(mock())
 
-        config.initKeyStores(cryptoService)
+        keyStoreHandler.initKeyStores()
 
         verify(signingSupplier).getOptional()
     }
@@ -102,7 +102,7 @@ class NodeKeyStoreUtilitiesTest {
         whenever(keySupplier.getOptional()).thenReturn(mock())
         whenever(signingSupplier.getOptional()).thenReturn(mock())
 
-        config.initKeyStores(bCryptoService)
+        KeyStoreHandler(config, bCryptoService).initKeyStores()
 
         verify(bCryptoService).resyncKeystore()
     }
@@ -118,6 +118,7 @@ class NodeKeyStoreUtilitiesTest {
     private val keySupplier = mock<FileBasedCertificateStoreSupplier>()
     private val trustRoot = mock<X509Certificate>()
     private val cryptoService = mock<CryptoService>()
+    private val keyStoreHandler = KeyStoreHandler(config, cryptoService)
 
     init {
         whenever(config.devMode).thenReturn(false)
